break components into files
