+++ x-pack/plugins/security_solution/public/timelines/components/timeline/header/save_timeline.test.tsx
@@ -6,212 +6,7 @@
 
 import React from 'react';
 import { shallow } from 'enzyme';
-import {
-  SaveTimelineButton,
-  SaveTimelineComponent,
-  TimelineTitleAndDescription,
-} from './save_timeline';
-import { useShallowEqualSelector } from '../../../../common/hooks/use_selector';
-import { useCreateTimelineButton } from '../properties/use_create_timeline';
-import { TimelineType } from '../../../../../common/types/timeline';
-import * as i18n from './translations';
-
-jest.mock('../../../../common/hooks/use_selector', () => ({
-  useShallowEqualSelector: jest.fn(),
-}));
-
-jest.mock('../../../../timelines/store/timeline', () => ({
-  timelineSelectors: {
-    selectTimeline: jest.fn(),
-  },
-}));
-
-jest.mock('../properties/use_create_timeline', () => ({
-  useCreateTimelineButton: jest.fn(),
-}));
-
-describe('TimelineTitleAndDescription', () => {
-  describe('save timeline', () => {
-    const props = {
-      timelineId: 'timeline-1',
-      toggleSaveTimeline: jest.fn(),
-      onSaveTimeline: jest.fn(),
-      updateTitle: jest.fn(),
-      updateDescription: jest.fn(),
-    };
-
-    const mockGetButton = jest.fn().mockReturnValue(<div data-test-subj="mock-discard-button" />);
-
-    beforeEach(() => {
-      (useShallowEqualSelector as jest.Mock).mockReturnValue({
-        description: '',
-        isSaving: true,
-        savedObjectId: null,
-        title: 'my timeline',
-        timelineType: TimelineType.default,
-      });
-      (useCreateTimelineButton as jest.Mock).mockReturnValue({
-        getButton: mockGetButton,
-      });
-    });
-
-    afterEach(() => {
-      (useShallowEqualSelector as jest.Mock).mockReset();
-      (useCreateTimelineButton as jest.Mock).mockReset();
-      mockGetButton.mockClear();
-    });
-
-    test('show proress bar while saving', () => {
-      const component = shallow(<TimelineTitleAndDescription {...props} />);
-      expect(component.find('[data-test-subj="progress-bar"]').exists()).toEqual(true);
-    });
-
-    test('Show correct header for save timeline modal', () => {
-      const component = shallow(<TimelineTitleAndDescription {...props} />);
-      expect(component.find('[data-test-subj="modal-header"]').prop('children')).toEqual(
-        i18n.SAVE_TIMELINE
-      );
-    });
-
-    test('Show correct header for save timeline template modal', () => {
-      (useShallowEqualSelector as jest.Mock).mockReturnValue({
-        description: '',
-        isSaving: true,
-        savedObjectId: null,
-        title: 'my timeline',
-        timelineType: TimelineType.template,
-      });
-      const component = shallow(<TimelineTitleAndDescription {...props} />);
-      expect(component.find('[data-test-subj="modal-header"]').prop('children')).toEqual(
-        i18n.SAVE_TIMELINE_TEMPLATE
-      );
-    });
-
-    test('Show name field', () => {
-      const component = shallow(<TimelineTitleAndDescription {...props} />);
-      expect(component.find('[data-test-subj="save-timeline-name"]').exists()).toEqual(true);
-    });
-
-    test('Show description field', () => {
-      const component = shallow(<TimelineTitleAndDescription {...props} />);
-      expect(component.find('[data-test-subj="save-timeline-description"]').exists()).toEqual(true);
-    });
-
-    test('Show discardTimelineButton', () => {
-      const component = shallow(<TimelineTitleAndDescription {...props} />);
-      expect(component.find('[data-test-subj="mock-discard-button"]').exists()).toEqual(true);
-    });
-
-    test('get discardTimelineButton with correct props', () => {
-      shallow(<TimelineTitleAndDescription {...props} />);
-      expect(mockGetButton).toBeCalledWith({
-        title: i18n.DISCARD_TIMELINE,
-        outline: true,
-        iconType: '',
-        fill: false,
-      });
-    });
-
-    test('get discardTimelineTemplateButton with correct props', () => {
-      (useShallowEqualSelector as jest.Mock).mockReturnValue({
-        description: 'xxxx',
-        isSaving: true,
-        savedObjectId: null,
-        title: 'my timeline',
-        timelineType: TimelineType.template,
-      });
-      shallow(<TimelineTitleAndDescription {...props} />);
-      expect(mockGetButton).toBeCalledWith({
-        title: i18n.DISCARD_TIMELINE_TEMPLATE,
-        outline: true,
-        iconType: '',
-        fill: false,
-      });
-    });
-
-    test('Show saveButton', () => {
-      const component = shallow(<TimelineTitleAndDescription {...props} />);
-      expect(component.find('[data-test-subj="save-button"]').exists()).toEqual(true);
-    });
-  });
-
-  describe('update timeline', () => {
-    const props = {
-      timelineId: 'timeline-1',
-      toggleSaveTimeline: jest.fn(),
-      onSaveTimeline: jest.fn(),
-      updateTitle: jest.fn(),
-      updateDescription: jest.fn(),
-    };
-
-    const mockGetButton = jest.fn().mockReturnValue(<div data-test-subj="mock-discard-button" />);
-
-    beforeEach(() => {
-      (useShallowEqualSelector as jest.Mock).mockReturnValue({
-        description: 'xxxx',
-        isSaving: true,
-        savedObjectId: '1234',
-        title: 'my timeline',
-        timelineType: TimelineType.default,
-      });
-      (useCreateTimelineButton as jest.Mock).mockReturnValue({
-        getButton: mockGetButton,
-      });
-    });
-
-    afterEach(() => {
-      (useShallowEqualSelector as jest.Mock).mockReset();
-      (useCreateTimelineButton as jest.Mock).mockReset();
-      mockGetButton.mockClear();
-    });
-
-    test('show proress bar while saving', () => {
-      const component = shallow(<TimelineTitleAndDescription {...props} />);
-      expect(component.find('[data-test-subj="progress-bar"]').exists()).toEqual(true);
-    });
-
-    test('Show correct header for save timeline modal', () => {
-      const component = shallow(<TimelineTitleAndDescription {...props} />);
-      expect(component.find('[data-test-subj="modal-header"]').prop('children')).toEqual(
-        i18n.NAME_TIMELINE
-      );
-    });
-
-    test('Show correct header for save timeline template modal', () => {
-      (useShallowEqualSelector as jest.Mock).mockReturnValue({
-        description: 'xxxx',
-        isSaving: true,
-        savedObjectId: '1234',
-        title: 'my timeline',
-        timelineType: TimelineType.template,
-      });
-      const component = shallow(<TimelineTitleAndDescription {...props} />);
-      expect(component.find('[data-test-subj="modal-header"]').prop('children')).toEqual(
-        i18n.NAME_TIMELINE_TEMPLATE
-      );
-    });
-
-    test('Show name field', () => {
-      const component = shallow(<TimelineTitleAndDescription {...props} />);
-      expect(component.find('[data-test-subj="save-timeline-name"]').exists()).toEqual(true);
-    });
-
-    test('Show description field', () => {
-      const component = shallow(<TimelineTitleAndDescription {...props} />);
-      expect(component.find('[data-test-subj="save-timeline-description"]').exists()).toEqual(true);
-    });
-
-    test('Show discardTimelineButton', () => {
-      const component = shallow(<TimelineTitleAndDescription {...props} />);
-      expect(component.find('[data-test-subj="mock-discard-button"]').exists()).toEqual(false);
-    });
-
-    test('Show saveButton', () => {
-      const component = shallow(<TimelineTitleAndDescription {...props} />);
-      expect(component.find('[data-test-subj="save-button"]').exists()).toEqual(true);
-    });
-  });
-});
+import { SaveTimelineComponent } from './save_timeline';
 
 describe('SaveTimelineComponent', () => {
   const props = {
@@ -241,30 +36,5 @@ describe('SaveTimelineComponent', () => {
   test('should not show a modal when showOverlay equals false', () => {
     const component = shallow(<SaveTimelineComponent {...props} />);
     expect(component.find('[data-test-subj="save-timeline-modal"]').exists()).toEqual(false);
-  });
-});
-
-describe('SaveTimelineButton', () => {
-  const props = {
-    timelineId: 'timeline-1',
-    showOverlay: false,
-    toolTip: 'tooltip message',
-    toggleSaveTimeline: jest.fn(),
-    onSaveTimeline: jest.fn(),
-    updateTitle: jest.fn(),
-    updateDescription: jest.fn(),
-  };
-  test('Show tooltip', () => {
-    const component = shallow(<SaveTimelineButton {...props} />);
-    expect(component.find('[data-test-subj="save-timeline-btn-tooltip"]').exists()).toEqual(true);
-  });
-
-  test('Hide tooltip', () => {
-    const testProps = {
-      ...props,
-      showOverlay: true,
-    };
-    const component = shallow(<SaveTimelineButton {...testProps} />);
-    expect(component.find('[data-test-subj="save-timeline-btn-tooltip"]').exists()).toEqual(false);
   });
 });
